+++ sample/src/components/WatchCase.vue
@@ -1,0 +1,7 @@
+<template>
+  <div :class="watchStyle.watch-target"></div>
+</template>
+
+<script setup lang="ts">
+import watchStyle from "../styles/watch-target.module.scss";
+</script>